factor out shared security rule
